+++ relay-event-normalization/src/normalize/span/mod.rs
@@ -1,10 +1,14 @@
 //! Span normalization logic.
 
 use regex::Regex;
+use relay_conventions::attributes::{
+    SENTRY__DSC__PROJECT_ID, SENTRY__DSC__TRACE_ID, SENTRY__DSC__TRANSACTION,
+};
 use relay_event_schema::protocol::{Event, SpanData, TraceContext};
 use relay_protocol::Annotated;
-use relay_sampling::DynamicSamplingContext;
 use std::sync::LazyLock;
+
+use crate::NormalizationConfig;
 
 pub mod ai;
 pub mod country_subregion;
@@ -23,6 +27,29 @@ pub static TABLE_NAME_REGEX: LazyLock<Regex> = LazyLock::new(|| {
     )
     .unwrap()
 });
+
+/// Applies [`relay_conventions`] configured normalizations to transaction spans.
+pub fn normalize_conventions(event: &mut Event) {
+    if let Some(data) = event
+        .contexts
+        .value_mut()
+        .as_mut()
+        .and_then(|c| c.get_mut::<TraceContext>())
+        .map(|c| &mut c.data)
+    {
+        crate::eap::normalize_attribute_names(data);
+    }
+
+    if let Some(spans) = event.spans.value_mut() {
+        for data in spans
+            .iter_mut()
+            .filter_map(|span| span.value_mut().as_mut())
+            .map(|span| &mut span.data)
+        {
+            crate::eap::normalize_attribute_names(data);
+        }
+    }
+}
 
 /// Replaces snake_case app start spans op with dot.case op.
 ///
@@ -60,14 +87,14 @@ pub fn normalize_app_start_spans(event: &mut Event) {
 ///
 /// If `sentry.dsc.trace_id` is already present in a span's `data`, the function does nothing for
 /// that span.
-pub fn normalize_dsc_for_event_spans(event: &mut Event, dsc: Option<&DynamicSamplingContext>) {
+pub fn normalize_dsc_for_event_spans(event: &mut Event, config: &NormalizationConfig) {
     if let Some(ctx) = event.context_mut::<TraceContext>() {
-        normalize_dsc_for_span_data(&mut ctx.data, dsc);
+        normalize_dsc_for_span_data(&mut ctx.data, config);
     }
     if let Some(spans) = event.spans.value_mut() {
         for span in spans {
             if let Some(span) = span.value_mut() {
-                normalize_dsc_for_span_data(&mut span.data, dsc);
+                normalize_dsc_for_span_data(&mut span.data, config);
             }
         }
     }
@@ -78,21 +105,27 @@ pub fn normalize_dsc_for_event_spans(event: &mut Event, dsc: Option<&DynamicSamp
 /// If `sentry.dsc.trace_id` is already present in `span_data`, the function does nothing.
 pub fn normalize_dsc_for_span_data(
     span_data: &mut Annotated<SpanData>,
-    dsc: Option<&DynamicSamplingContext>,
+    config: &NormalizationConfig,
 ) {
-    let Some(dsc) = dsc else {
+    let Some(dsc) = config.dsc else {
         return;
     };
 
     let data = span_data.get_or_insert_with(SpanData::default);
-    if data.sentry_dsc_trace_id.value().is_some() {
+    if data.get_value(SENTRY__DSC__TRACE_ID).is_some() {
         return;
     }
-    data.sentry_dsc_trace_id = Annotated::new(dsc.trace_id.to_string());
+    data.insert_value(SENTRY__DSC__TRACE_ID, dsc.trace_id.to_string());
     if let Some(project_id) = &dsc.project_id {
-        data.sentry_dsc_project_id = Annotated::new(project_id.to_string());
+        data.insert_value(SENTRY__DSC__PROJECT_ID, project_id.to_string());
     }
     if let Some(transaction) = &dsc.transaction {
-        data.sentry_dsc_transaction = Annotated::new(transaction.to_string());
+        // To match the behaviour of the `count_per_root` metric, which had its tags
+        // removed if they were over the limit.
+        match transaction.len() <= config.max_tag_value_length {
+            true => data.insert_value(SENTRY__DSC__TRANSACTION, transaction.clone()),
+            // Keep an empty value around for the DS job
+            false => data.insert_value(SENTRY__DSC__TRANSACTION, "".to_owned()),
+        }
     }
 }