+++ relay-server/src/services/objectstore.rs
@@ -50,7 +50,7 @@ pub enum Objectstore {
     TraceAttachment(Managed<StoreTraceAttachment>),
     EventAttachment(Managed<StoreAttachment>),
     RawProfile(Managed<StoreRawProfile>),
-    Create(Create, Sender<Result<UploadRef, Error>>),
+    Create(CreateMultipart, Sender<Result<UploadRef, Error>>),
     Stream(Stream, Sender<Result<ObjectstoreKey, Error>>),
 }
 
@@ -137,7 +137,7 @@ impl MessageKind {
 }
 
 /// A request to create a new objectstore multipart upload.
-pub struct Create {
+pub struct CreateMultipart {
     /// The sentry org.
     pub organization_id: OrganizationId,
     /// The sentry project.
@@ -146,10 +146,10 @@ pub struct Create {
     pub key: String,
 }
 
-impl FromMessage<Create> for Objectstore {
+impl FromMessage<CreateMultipart> for Objectstore {
     type Response = AsyncResponse<Result<UploadRef, Error>>;
 
-    fn from_message(message: Create, sender: Sender<Result<UploadRef, Error>>) -> Self {
+    fn from_message(message: CreateMultipart, sender: Sender<Result<UploadRef, Error>>) -> Self {
         Self::Create(message, sender)
     }
 }
@@ -768,8 +768,8 @@ impl ObjectstoreServiceInner {
         Ok(Some(stored_key))
     }
 
-    async fn handle_create(&self, create: Create) -> Result<UploadRef, Error> {
-        let Create {
+    async fn handle_create(&self, create: CreateMultipart) -> Result<UploadRef, Error> {
+        let CreateMultipart {
             organization_id,
             project_id,
             key,
@@ -783,6 +783,7 @@ impl ObjectstoreServiceInner {
             .send()
             .await?;
         debug_assert_eq!(&key, multipart_upload.key());
+
         let upload_id = multipart_upload.upload_id();
 
         Ok(UploadRef {
@@ -947,9 +948,27 @@ impl ObjectstoreServiceInner {
                     while let Some((i, chunk)) = body.next().await {
                         let chunk = chunk?;
                         let part_number = u32::try_from(i + 1)
-                            .map_err(|_| objectstore_client::Error::InvalidPartNumber(u32::MAX))?;
-                        let part = multipart_upload.put(chunk, part_number, None).await?;
-                        parts.push(part);
+                        .map_err(|_| objectstore_client::Error::InvalidPartNumber(u32::MAX))?;
+                        relay_log::trace!("Part number {part_number}");
+
+                        // NOTE: This is a retry loop within a retry loop (see caller of this function).
+                        // if we keep the Rechunked approach we might as well remove the outer loop for streaming uploads.
+                        let mut attempts = 0;
+                        let part = loop {
+                            let result = multipart_upload.put(chunk.clone(), part_number, None).await;
+                            attempts += 1;
+                            if attempts < self.max_attempts.get()
+                                && matches!(&result, Err(e) if is_retryable(e))
+                            {
+                                relay_log::trace!("Attempt {attempts}: Failed with {result:?}, retrying");
+                                tokio::time::sleep(self.retry_interval).await;
+                            } else {
+                                relay_log::trace!("Final attempt");
+                                break result;
+                            }
+                        };
+
+                        parts.push(part?);
                     }
                     multipart_upload.complete(parts).await?
                 });