+++ relay-server/src/statsd.rs
@@ -747,6 +747,12 @@ pub enum RelayCounters {
     ///  - `sdk`: The name of the Sentry SDK sending the envelope. This tag is only set for
     ///    Sentry's SDKs and defaults to "proprietary".
     EnvelopeItemBytes,
+    /// Number of envelopes rejected because of size limits.
+    ///
+    /// This metric is tagged with:
+    ///  - `item`: The type of the items being counted.
+    ///  - `limit`: Which limit was breached.
+    EnvelopeSizeLimited,
     /// Number of times an envelope from the buffer is trying to be popped.
     BufferTryPop,
     /// Number of envelopes spool to disk.
@@ -1029,6 +1035,7 @@ impl CounterMetric for RelayCounters {
             RelayCounters::EnvelopeRejected => "event.rejected",
             RelayCounters::EnvelopeItems => "event.items",
             RelayCounters::EnvelopeItemBytes => "event.item_bytes",
+            RelayCounters::EnvelopeSizeLimited => "envelope.rejected.size",
             RelayCounters::BufferTryPop => "buffer.try_pop",
             RelayCounters::BufferSpooledEnvelopes => "buffer.spooled_envelopes",
             RelayCounters::BufferUnspooledEnvelopes => "buffer.unspooled_envelopes",