+++ relay-server/src/utils/native.rs
@@ -10,7 +10,6 @@ use chrono::{TimeZone, Utc};
 use minidump::{
     MinidumpAnnotation, MinidumpCrashpadInfo, MinidumpModuleList, Module, StabilityReport,
 };
-use relay_dynamic_config::Feature;
 use relay_event_schema::protocol::{
     ClientSdkInfo, Context, Contexts, Event, Exception, JsonLenientString, Level, Mechanism,
     StabilityReportContext, Values,
@@ -18,7 +17,6 @@ use relay_event_schema::protocol::{
 use relay_protocol::{Annotated, Value, get_value};
 
 use crate::envelope::{Item, ItemType};
-use crate::services::projects::project::ProjectInfo;
 
 type Minidump<'a> = minidump::Minidump<'a, &'a [u8]>;
 
@@ -40,6 +38,13 @@ struct NativePlaceholder {
     mechanism_type: &'static str,
 }
 
+#[derive(Clone, Copy)]
+/// What to do with additional exceptions in a minidump / apple crash report event.
+pub enum AdditionalExceptions {
+    Retain,
+    Delete,
+}
+
 /// Writes a placeholder to indicate that this event has an associated minidump or an apple
 /// crash report.
 ///
@@ -48,7 +53,7 @@ struct NativePlaceholder {
 fn write_native_placeholder(
     event: &mut Event,
     placeholder: NativePlaceholder,
-    project_info: &ProjectInfo,
+    additional_exceptions: AdditionalExceptions,
 ) {
     // Events must be native platform.
     let platform = event.platform.value_mut();
@@ -71,7 +76,6 @@ fn write_native_placeholder(
         .value_mut()
         .get_or_insert_with(Vec::new);
 
-    let allow_multiple_exceptions = project_info.has_feature(Feature::MinidumpMultiException);
     if let Some(exc) = exceptions.first() {
         relay_log::info!(
             additional_exceptions = exceptions.len(),
@@ -79,13 +83,12 @@ fn write_native_placeholder(
             additional_exception_mechanism = ?get_value!(exc.mechanism.ty),
             sentry_project = ?event.project,
             event_id = ?event.id,
-            has_feature = allow_multiple_exceptions,
             platform = ?event.platform,
             "Native event has additional exceptions",
         )
     }
 
-    if !allow_multiple_exceptions {
+    if matches!(additional_exceptions, AdditionalExceptions::Delete) {
         exceptions.clear(); // clear previous errors if any
     }
 
@@ -223,14 +226,18 @@ fn write_crashpad_annotations(
 ///
 /// This function operates at best-effort. It always attaches the placeholder and returns
 /// successfully, even if the minidump or part of its data cannot be parsed.
-pub fn process_minidump(event: &mut Event, item: &Item, project_info: &ProjectInfo) {
+pub fn process_minidump(
+    event: &mut Event,
+    item: &Item,
+    additional_exceptions: AdditionalExceptions,
+) {
     debug_assert_eq!(item.ty(), &ItemType::Attachment);
     let placeholder = NativePlaceholder {
         exception_type: "Minidump",
         exception_value: "Invalid Minidump",
         mechanism_type: "minidump",
     };
-    write_native_placeholder(event, placeholder, project_info);
+    write_native_placeholder(event, placeholder, additional_exceptions);
 
     if item.is_attachment_ref() {
         // We don't have a full minidump, just a placeholder for something that was uploaded
@@ -293,11 +300,11 @@ pub fn process_minidump(event: &mut Event, item: &Item, project_info: &ProjectIn
 
 /// Writes minimal information into the event to indicate it is associated with an Apple Crash
 /// Report.
-pub fn process_apple_crash_report(event: &mut Event, project_info: &ProjectInfo) {
+pub fn process_apple_crash_report(event: &mut Event, additional_exceptions: AdditionalExceptions) {
     let placeholder = NativePlaceholder {
         exception_type: "AppleCrashReport",
         exception_value: "Invalid Apple Crash Report",
         mechanism_type: "applecrashreport",
     };
-    write_native_placeholder(event, placeholder, project_info);
+    write_native_placeholder(event, placeholder, additional_exceptions);
 }