+++ tests/integration/test_nel.py
@@ -74,6 +74,7 @@ def test_nel_converted_to_logs(mini_sentry, relay):
                         "type": "string",
                         "value": time_within_delta(expect_resolution="ns"),
                     },
+                    "sentry.relay.ingress": {"type": "string", "value": "integration"},
                 },
                 "body": "The user agent successfully received a response, but it had a 500 status code",
                 "level": "warn",